+++ src/DistanceType.php
@@ -2,15 +2,14 @@
 
 namespace Ngt;
 
-// TODO use enum when PHP 8.0 reaches EOL
-class DistanceType
+enum DistanceType
 {
-    public const L1 = 0;
-    public const L2 = 1;
-    public const Hamming = 2;
-    public const Angle = 3;
-    public const Cosine = 4;
-    public const NormalizedAngle = 5;
-    public const NormalizedCosine = 6;
-    public const Jaccard = 7;
+    case L1;
+    case L2;
+    case Hamming;
+    case Angle;
+    case Cosine;
+    case NormalizedAngle;
+    case NormalizedCosine;
+    case Jaccard;
 }