+++ src/ObjectType.php
@@ -2,10 +2,9 @@
 
 namespace Ngt;
 
-// TODO use enum when PHP 8.0 reaches EOL
-class ObjectType
+enum ObjectType
 {
-    public const Float = 0;
-    public const Float16 = 1;
-    public const Integer = 2;
+    case Float;
+    case Float16;
+    case Integer;
 }